refactor(runtime): regroup elicitation and resume code into focused files
runtime.go was a 1081-line catch-all that mixed five distinct concerns: the LocalRuntime struct + Opts + constructor, current-agent helpers, MCP prompts, session/permissions/lifecycle plumbing, the Resume API surface, and the elicitation API + handler. Splitting the elicitation and resume concerns into their own files removes \u003c100 lines from runtime.go and groups related logic by topic.

Move ElicitationResult / ElicitationError / ElicitationRequestHandler / ResumeElicitation method / elicitationHandler method, plus the elicitationBridge type from elicitation_bridge.go, into a single elicitation.go. Everything elicitation-related now lives in one ~140-line file.

Move ResumeType + constants + ResumeRequest + ResumeApprove* helpers + IsValidResumeType / ValidResumeTypes (previously split across runtime.go and resume_validation.go) into resume.go. Doc comments expanded to make the type act as an entry point for readers.

Test files renamed correspondingly: resume_validation_test.go \u2192 resume_test.go, elicitation_bridge_test.go \u2192 elicitation_test.go (TestElicitationError_Error moved over too).

Net effect: runtime.go drops from 1081 to 963 lines; the package gains two well-named topical files; no behaviour changes.
