--- pkg/runtime/resume_validation.go
@@ -1,24 +1,0 @@
-package runtime
-
-// IsValidResumeType validates confirmation values coming from /resume
-func IsValidResumeType(t ResumeType) bool {
-	switch t {
-	case ResumeTypeApprove,
-		ResumeTypeApproveSession,
-		ResumeTypeApproveTool,
-		ResumeTypeReject:
-		return true
-	default:
-		return false
-	}
-}
-
-// ValidResumeTypes returns all allowed confirmation values
-func ValidResumeTypes() []ResumeType {
-	return []ResumeType{
-		ResumeTypeApprove,
-		ResumeTypeApproveSession,
-		ResumeTypeApproveTool,
-		ResumeTypeReject,
-	}
-}
